Rename RPC module to YARPC
Fixes #267
