+++ modules/yarpc/thriftrw-plugin-thriftsync/doc.go
@@ -34,7 +34,7 @@
 //
 // • Install thriftsync from vendor
 //
-// go install ./vendor/go.uber.org/fx/modules/rpc/thriftrw-plugin-thriftsync
+// go install ./vendor/go.uber.org/fx/modules/yarpc/thriftrw-plugin-thriftsync
 //
 // • Run thriftrw code genration with thriftsync
 //
@@ -49,7 +49,7 @@
 //     go install ./vendor/go.uber.org/thriftrw
 //
 //     @echo "Installing thriftrw-plugin-thriftsync..."
-//     go install ./vendor/go.uber.org/fx/modules/rpc/thriftrw-plugin-thriftsync
+//     go install ./vendor/go.uber.org/fx/modules/yarpc/thriftrw-plugin-thriftsync
 //
 //   thriftsync: deps
 //     thriftrw --plugin="thriftsync" <thrift filepath>
