+++ service/README.md
@@ -14,7 +14,7 @@ package main
 
 import (
   "go.uber.org/fx/config"
-  "go.uber.org/fx/modules/rpc"
+  "go.uber.org/fx/modules/yarpc"
   "go.uber.org/fx/service"
 )
 
@@ -24,7 +24,7 @@ func main() {
     // The list of module creators for this service, in this case
     // creates a Thrift RPC module called "keyvalue"
     "keyvalue",
-    rpc.ThriftModule(rpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
+    yarpc.ThriftModule(yarpc.CreateThriftServiceFunc(NewYarpcThriftHandler)),
   ).Build()
 
   if err != nil {
